requirements.txt
torch>=1.11.0
torchvision>=0.12.0
numpy>=1.22.3
matplotlib>=3.5.1
tqdm>=4.64.0
pandas>=1.4.2
tensorboard>=1.14
Pillow>=9.1.0
pyyaml
sklearn 
h5py 
lightning
easydict